undo finalization of terminal completion provider API (#276174)
addresses #276142 in main
