+++ extensions/terminal-suggest/package.json
@@ -14,6 +14,7 @@
     "Other"
   ],
   "enabledApiProposals": [
+    "terminalCompletionProvider",
     "terminalShellEnv"
   ],
   "contributes": {
